HLT-vkl-26/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
pymysql==1.1.1
pydantic-settings==2.9.1
